moving tests to proper place
